+++ OUDS/Core/Tokens/SemanticTokens/Sources/Values/ElevationSemanticTokens.swift
@@ -53,7 +53,6 @@ public protocol ElevationSemanticTokens {
     var elevationBlurNone: ElevationBlurSemanticToken { get }
     var elevationBlurRaised: ElevationBlurSemanticToken { get }
     var elevationBlurDrag: ElevationBlurSemanticToken { get }
-    var elevationBlurOverlayDefault: ElevationBlurSemanticToken { get }
     var elevationBlurOverlayEmphasized: ElevationBlurSemanticToken { get }
     var elevationBlurStickyDefault: ElevationBlurSemanticToken { get }
     var elevationBlurStickyEmphasized: ElevationBlurSemanticToken { get }